backend/pyproject.toml
[project]
name = "etl-qa-agent"
version = "0.1.0"
description = "Agente QA per ETL BigQuery - singolo agente ADK/Gemini"
requires-python = ">=3.11"

dependencies = [
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",

    # Google ADK + Vertex AI
    "google-adk>=0.4.0",
    "google-cloud-aiplatform>=1.70.0",

    # Data stores
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-firestore>=2.19.0",

    # GitHub
    "PyGithub>=2.4.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
